AppointmeAvaliables en el Backend

+++ backend/culturefit/src/main/java/com/culturefit/culturefit/domains/User.java
@@ -56,5 +56,5 @@ public class User {
     @Column( unique = true)
     private String dni;
 
-    private Long appointmentsAvailables = 0L;
+    private int appointmentsAvailables = 0;
 }
